requirements.txt
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
PyYAML==3.13
Werkzeug==0.14.1
beautifulsoup4==4.6.0
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
google-api-core==1.2.1
google-auth==1.5.0
google-cloud-language==1.0.2
googleapis-common-protos==1.6.0b4
grpcio==1.13.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
jieba3k==0.35.1
lxml==4.2.3
newspaper3k==0.2.6
nltk==3.3
protobuf==3.6.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
requests-file==1.4.3
rsa==3.4.2
six==1.11.0
tldextract==2.2.0
urllib3==1.23
wheel==0.26.0
